+++ src/shared/AppRoutes.ts
@@ -1,0 +1,7 @@
+export enum AppRoutes {
+    Main = '/',
+    Offer = '/offer',
+    Favorites = '/favorites',
+    Login = '/login'
+}
+
