+++ src/shared/AutorizationStatus.ts
@@ -1,0 +1,5 @@
+export enum AutorizationStatus {
+    Auth = 'AUTH',
+    NoAuth = 'NO_AUTH',
+    Unkhown = 'ANKHNOWN'
+}
