--- src/shared/constats.ts
@@ -1,12 +1,0 @@
-export enum AppRoutes {
-  Main = '/',
-  Offer = '/offer',
-  Favorites = '/favorites',
-  Login = '/login',
-}
-
-export enum AutorizationStatus {
-  Auth = 'AUTH',
-  NoAuth = 'NO_AUTH',
-  Unkhown = 'ANKHNOWN'
-}
