+++ tsconfig.json
@@ -1,13 +1,13 @@
 {
   "compilerOptions": {
     "target": "ES2020",
-    "lib": ["DOM", "DOM.Iterable", "ESNext"],
+    "lib": [ "DOM", "DOM.Iterable", "ESNext" ],
     "module": "ESNext",
     "skipLibCheck": true,
-    "types": ["vitest/globals", "@testing-library/jest-dom"],
+    "types": [ "vitest/globals", "@testing-library/jest-dom" ],
     "baseUrl": ".",
     "paths": {
-      "src/*":["./src/*"]
+      "src/*": [ "./src/*" ]
     },
 
     /* Bundler mode */
@@ -23,8 +23,8 @@
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true,
-    "forceConsistentCasingInFileNames": true,
+    "forceConsistentCasingInFileNames": true
   },
-  "include": ["src", "vite.config.ts"],
-  "references": [{ "path": "./tsconfig.node.json" }]
+  "include": [ "src", "vite.config.ts" ],
+  "references": [ { "path": "./tsconfig.node.json" } ]
 }
